// src/types/ItemType.ts
export interface ItemType {
  _id?: string
  group?: string
  subGroup?: string
  category?: string
  brand?: string
  code?: string
  name?: string
  description?: string
  manufacturerCode?: string
  barcode?: string
  vatRate?: number
  withHoldingTaxRate?: number
  unit?: string
  lastModified?: string
  passive?: boolean
}